% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2023-04-24_DisclosuR.R
\name{newswire_segmenter_folder}
\alias{newswire_segmenter_folder}
\title{Newswire segmenter (multiple files)}
\usage{
newswire_segmenter_folder(
  folder_path,
  sentiment = FALSE,
  emotion = FALSE,
  regulatory_focus = FALSE,
  laughter = FALSE,
  narcissism = FALSE,
  text_clustering = FALSE
)
}
\arguments{
\item{sentiment}{Performs dictionary-based sentiment analysis
based on the \code{\link[SentimentAnalysis]{analyzeSentiment}}
function (default: FALSE)}

\item{emotion}{Performs dictionary-based emotion analysis based on the
\code{\link[syuzhet]{get_nrc_sentiment}} function (default: FALSE)}

\item{regulatory_focus}{Calculates the number of words indicative for
promotion and prevention focus based on the
dictionary developed by \href{https://www.jstor.org/stable/pdf/43589392.pdf?casa_token=Zye71gMR5RoAAAAA:G76SwrZDfECXcyMeROZkhSgBsvGWyiULh-Fy7kAb1fpxS4DCxpZZbocUfTesOm90lAa5vl-lKOvYVI7ZKxNS9eP0MlVN9eRv31kMGhSPD6ldjzqrSyeF)}{Gamache et al., 2015}
(default: FALSE)}

\item{laughter}{Counts the number of times laughter was indicated in a quote. (default: FALSE)}

\item{narcissism}{Counts the number of pronoun usage and calculates the ratio of first-person singular to
first-person plural pronouns. This measure is derived from \href{https://journals.sagepub.com/doi/pdf/10.1177/0001839214554989?casa_token=SE3k4io6um0AAAAA:gpJvbVmEkbA1-b9J5lSRvRv53WrAvWBWgqSK74s8AMhu3MCq9hy1mUkXpS3MM3XHUf8ZpJTNbGWKnQ}{Zhu & Chen, (2015}
(default: FALSE)}

\item{file}{The name of the PDF file which the data are to be read
from. If it does not contain an absolute path, the file name is
relative to the current working directory, getwd().}

\item{title_clustering}{Applies a document categorization using a dictionary developed based on the framework developed by \href{https://journals.aom.org/doi/abs/10.5465/amj.2013.0288}{Graffin et al., 2016}. (default: FALSE)}
}
\value{
An R data frame with each row representing one newswire article. The columns indicate the title, text,
newswire, date, and weekday. (default: FAKSE)

An R data frame with each row representing one newswire article. The columns indicate the title, text,
newswire, date, and weekday. Depending on the additional arguemnts, the output data can also
contain sentiment, emotion, regulatory focus, laughter, narcissism and text cluster based on the Graffin et al.
categories.
}
\description{
Takes all PDF documents in a folder containing news-wire documents obtained from NexisUni and transforms them into
an R data frame consisting of one row per document.
}
\examples{
newswire_df <- newswire_segmenter(file = "press_release.pdf");
newswire_df_sentiment <- newswire_segmenter(file = "press_release.pdf", sentiment = TRUE);
}
